+++ CHANGELOG.md
@@ -54,6 +54,11 @@ To install release candidates run `[sudo] gem install cocoapods --pre`
 * Cache globbing in `PathList` to speed up `pod install`.
   [Vincent Isambart](https://github.com/vincentisambart)
 
+* CocoaPods will validate your podfile and try to identify problems 
+  and conflicts in how you've specified the dependencies
+  [Hugo Tunius](https://github.com/k0nserv)
+  [#995](https://github.com/CocoaPods/CocoaPods/issues/995)
+
 ##### Bug Fixes
 
 * Added recursive support to the public headers of vendored frameworks